Change price datatype from float to BigDecimal.

+++ src/main/kotlin/com/noorifytech/maha/dto/Price.kt
@@ -1,3 +1,5 @@
 package com.noorifytech.maha.dto
 
-data class Price(val price: Float)
+import java.math.BigDecimal
+
+data class Price(val price: BigDecimal)
